refactor(store radio button card): rename files to use kebab case
